backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
whitenoise>=6.6
cryptography>=42.0
gunicorn>=20.1.0